# packages/sdk/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me/layouts/{name}": {
      "put": {
        "operationId": "LayoutsController_save",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "layout": {
                    "type": "object",
                    "propertyNames": {
                      "type": "string"
                    },
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "layout"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Save (create or replace) a named layout.",
        "tags": [
          "terminal"
        ]
      },
      "delete": {
        "operationId": "LayoutsController_remove",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete a named layout.",
        "tags": [
          "terminal"
        ]
      }
    },
...
    "/price-alerts": {
      "get": {
        "operationId": "AlertsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "active",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "The user’s price and indicator alerts, newest first.",
        "tags": [
          "terminal"
        ]
      },
      "post": {
        "operationId": "AlertsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "symbol": {
                    "type": "string",
                    "pattern": "^[A-Z0-9][A-Z0-9._-]{0,31}$"
                  },
                  "condition": {
                    "type": "string",
                    "enum": [
                      "price_above",
                      "price_below",
                      "rsi_above",
                      "rsi_below"
                    ]
                  },
                  "threshold": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "timeframe": {
                    "type": "string",
                    "enum": [
                      "1m",
                      "5m",
                      "15m",
                      "1h",
                      "4h",
                      "1D"
                    ]
                  },
                  "note": {
                    "type": "string",
                    "maxLength": 120
                  }
                },
                "required": [
                  "symbol",
                  "condition",
                  "threshold"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a price (above/below) or RSI(14) alert on a timeframe. Audited as alert.created.",
        "tags": [
          "terminal"
        ]
      }
    },
    "/price-alerts/{id}": {
      "delete": {
        "operationId": "AlertsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Cancel an active alert. Audited as alert.cancelled.",
        "tags": [
          "terminal"
        ]
      }
    },
...
  },
  "info": {
    "title": "KORA API",
    "description": "KORA platform API (PAPER only). Money and prices are decimal strings.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kora_at"
      }
    },
...
  }
}